requirements.txt
flask==2.3.3
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==0.8.11
werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
